+++ packages/credential_manager_ios/lib/src/credential_manager_ios_plugin.dart
@@ -4,6 +4,10 @@ import 'package:flutter/foundation.dart';
 
 /// iOS plugin registration
 class CredentialManagerIosPlugin {
+  /// Prevents instantiation. This plugin is only used for its static
+  /// [registerWith] method during plugin registration.
+  const CredentialManagerIosPlugin._();
+
   /// Registers the iOS implementation
   static void registerWith() {
     if (defaultTargetPlatform == TargetPlatform.iOS) {